refactor(sdk)!: move worker-connection-manager types to the helpers submodule
AuthInput, AuthResult, and the six On*Registration{Input,Result} types now live
only in @iii-dev/helpers/worker-connection-manager and are removed from the root
SDK public surface (no re-export). MiddlewareFunctionInput stays in the root SDK
because it references the core TriggerAction type.
